a5 wiring: move tensormap + fanout++ to wiring thread (step 4)
Pull sync_tensormap / compute_task_fanin / register_task_outputs and the
producer fanout_count++ loop out of the orchestrator hot path and run
them on the dedicated wiring thread. orch now only allocs the slot,
copies tensor/scalar to payload, resolves explicit_deps into the inline
fanin prefix, and pushes a TaskSubmit event.

- New scheduler_wire.cpp owns wire_task + drain_wiring_queue (needs the
  full PTO2OrchestratorState definition for tensor_map / fanin_pool, so
  it can't live in the scheduler header).
- TensorArgType is now `enum class : uint8_t`; tags + in_manual_scope
  ride on PTO2WiringEvent (40B) instead of PTO2TaskPayload. tag copy is
  a single 16B memcpy in for_task_submit.
- sync_tensormap moves from drain entry to advance_ring_pointers (event-
  driven: only fires when last_task_alive advances).
- advance_lock CAS dropped — wiring is the single writer after step 1.
- wait_for_tensor_ready takes a drain+idle barrier before reading
  tensor_map; skipped when the looked-up tensor has no owner.
- PTO2FaninBuilder moved to its own header with an indirect inline_slots
  pointer so orch's explicit_dep prefix and wiring's tensormap lookup
  share payload.fanin_inline_slot_states[] (no 512B copy between them).
- aicpu_executor logs wire_cost in parallel with orch_cost / sched_cost.

paged_attention_unroll Case1 (100 rounds, trim k=15, device 2, ORCH_
PROFILING off): orch_cost 815.50us, sched_max 1084.38us, wire_cost
1243.62us. Step 1+2 baseline d50cc0b3 was 843.40 / 1078.21.

+++ src/common/task_interface/tensor_arg.h
@@ -52,8 +52,12 @@ static_assert(
 
 /**
  * TensorArgType - Distinguishes inputs, outputs, and in-place updates
+ *
+ * Stored as uint8_t so a full MAX_TENSOR_ARGS=16-slot array fits in a single
+ * 16-byte chunk. Existing AICPU code paths (dep_gen, etc.) already treat the
+ * tag as a raw byte; this just makes the canonical width match.
  */
-enum class TensorArgType : int32_t {
+enum class TensorArgType : uint8_t {
     INPUT = 0,            // Read-only input buffer
     OUTPUT = 1,           // Write-only output buffer (runtime allocates)
     INOUT = 2,            // Read-then-write: modifier for downstream
